fix(db): Correct Chat entity parameter name
This commit fixes a bug in the `ChatRepository` where an incorrect parameter name was used when creating a new `Chat` entity.

The `Chat` constructor was being called with the `name` parameter, but it was expecting `chatName`. This has been corrected to use `chatName`, ensuring messages are inserted into the correct chat.